Add component to handle image uploading and saving
